Object refactor
