+++ Makefile
@@ -10,13 +10,13 @@ calc: build build/calc_unbounded_int.o build/unbounded_int.o
 test: build build/test_unbounded.o build/unbounded_int.o
 	$(CC) $(CFLAGS) $(LDLIBS)  build/unbounded_int.o build/test_unbounded.o -o $@
 
-build/test_unbounded.o: build/unbounded_int.o unbounded_int.h
+build/test_unbounded.o: build/unbounded_int.o src/unbounded_int.h
 	$(CC) $(CFLAGS) $(LDLIBS)  -c test_unbounded.c -o $@
 
-build/unbounded_int.o: unbounded_int.c unbounded_int.h
+build/unbounded_int.o: src/unbounded_int.c src/unbounded_int.h
 	$(CC) $(CFLAGS) $(LDLIBS) -c unbounded_int.c -o $@
 
-build/calc_unbounded_int.o: calc_unbounded_int.c unbounded_int.h
+build/calc_unbounded_int.o: src/calc_unbounded_int.c src/unbounded_int.h
 	$(CC) $(CFLAGS) $(LDLIBS) -c calc_unbounded_int.c -o $@
