+++ src/calc_unbounded_int.c
@@ -13,20 +13,13 @@
 #include <errno.h>
 
 #include "unbounded_int.h"
-
-/**
- * Parser and lexer errors.
- */
-enum ERROR {
-    INVALID_OPERATOR, INVALID_SYNTAX, MISSING_BLANK, UNKNOWN_CHARACTER, INTERNAL, TOO_MANY_ARGUMENTS, MISSING_ARGUMENTS
-};
-
-/**
- * Print on the stderr the corresponding message according to the code given in argument.
- * @param error The error's code.
- * @return The error's message.
- */
-static char *error_getMessage(enum ERROR error);
+#include "lib.h"
+#include "hashmap.h"
+#include "execerror.h"
+#include "function.h"
+#include "token.h"
+#include "ast.h"
+#include "buffer.h"
 
 /**
  * Name of the open file.
@@ -40,105 +33,8 @@ static FILE *OUT;
 static int EXIT_REQUEST = 0;
 static int CALL_ID = 0;
 
-#define pERROR(error)(fprintf(stderr, "%s in file %s in line %d\n", error_getMessage(error), FILE_NAME, FILE_LINE))
-#define printErr(c)(fprintf(stderr, "%s in file %s in line %d\n %s\n", strerror(errno), __FILE__, __LINE__, (c)))
-#define DEFAULT_OP '\0'
-#define MULTIPLICATION(a, b)(unbounded_int_produit(a,b))
-#define ADDITION(a, b)(unbounded_int_somme(a,b))
-#define SUBSTRACTION(a, b)(unbounded_int_difference(a,b))
-//#define DIVISION(a, b)(unbounded_int_division(a, b))
-#define EQUALS(t, a, b)(tree_add((t), (a), (b)))
-//#define MODULO(a, b)((a) % (b))
-
 #define BUFFER_SIZE 16
-#define HASHMAP_INITIAL_SIZE 11
-#define HASHMAP_MIN_RATIO ((float)0.25)
-#define HASHMAP_MAX_RATIO ((float)0.75)
-#define HASH_A ((sqrt(5)-1)/2)
 #define INCR_CALL (CALL_ID = CALL_ID + 19 % INT_MAX)
-
-
-
-
-
-/* #####################################################################################################################
- * Token's functions.
- */
-
-/**
-* Token types.
-*/
-typedef enum TokenType {
-    NUMBER, VAR, OPERATOR, VOID, FUNCTION
-} TokenType;
-
-/**
- * Lexers Tokens.
- */
-typedef struct Token {
-    char *data;
-    enum TokenType type;
-} Token;
-
-/**
- * Free a token.
- * @param t The token which will be free
- */
-static void token_free(Token t);
-
-/**
- * Create a new token according to the string given in argument and the type.
- * @param s The token' data.
- * @param type The token's type.
- * @return The created Token.
- */
-static Token token_new(char *s, size_t len, enum TokenType type);
-
-
-
-
-/* #####################################################################################################################
- * Parser's buffer's functions.
-*/
-
-/**
- * Parser Buffer.
-*/
-typedef struct Buffer {
-    char *buffer;
-    int capacity;
-    int length;
-} Buffer;
-
-/**
- * Create a dynamic buffer.
- * @return The buffer pointer.
- */
-static Buffer *buffer_new();
-
-/**
- * Clear the buffer and puts its capacity to the BUFFER_SIZE.
- * @param buffer The buffer pointer.
- * @return the buffer if everything went fine, NULL otherwise
- */
-static Buffer *buffer_clear(Buffer *buffer);
-
-/**
- * Free the buffer.
- *
- * @param buffer The buffer which will be free.
- * @return NULL.
- */
-static Buffer *buffer_free(Buffer *buffer);
-
-/**
- * Add the given element in the given buffer
- * @param buffer The buffer
- * @param e The element which will be added to the buffer
- * @return 1 if the element is added to the buffer, 0 otherwise
- */
-static int buffer_add(Buffer *buffer, char e);
-
 
 /* ####################################################################################################################
  * STD functions
@@ -241,183 +137,6 @@ static int node_add(Node **pNode, unbounded_int n, const char *string);
  */
 static unbounded_int node_getValue(Node **node, char *string);
 
-
-/*#####################################################################################################################
- * Function's functions
- */
-typedef enum FuncType {
-    STD, DUMMY, NONE, CALL
-} FuncType;
-
-typedef enum RetType {
-    NUM_TYPE, VOID_TYPE
-} RetType;
-
-
-typedef struct Function Function;
-
-struct Function {
-    char *name;
-    unsigned short requested;
-    unsigned short argc;
-    RetType retType;
-    unbounded_int *argv;
-    char **argn;
-
-    int (*func)(int, unbounded_int *, char **);
-};
-
-static Function
-function_new(char *name, RetType type, int (*function)(int, unbounded_int *, char **),
-             unsigned short requestedArguments);
-
-static void function_free(Function f);
-
-#define FUNCTION_NULL (Function) {.name = NULL, .requested = 0, .argc = 0, .retType= VOID_TYPE, .argv = NULL, .argn = NULL}
-
-/* #####################################################################################################################
-* funcHashMap's declaration
-*/
-#define DUMMY_DATA(h) ((HashMapData) {.hash = (h), .function = FUNCTION_NULL, .type=DUMMY})
-#define NONE_DATA ((HashMapData) {.hash = 0,.function = FUNCTION_NULL, .type=NONE})
-
-typedef struct HashMap HashMap;
-typedef struct HashMapData HashMapData;
-
-struct HashMapData {
-    long long hash;
-    Function function;
-    FuncType type;
-};
-struct HashMap {
-    HashMapData *data;
-    size_t capacity;
-    size_t keyNumber;
-    size_t dummyNumber;
-    float minRatio, maxRatio;
-};
-
-static void resize(HashMap *map, size_t new);
-
-static HashMapData hashMap_get(HashMap *map, const char *name);
-
-static HashMap *HashMap_new();
-
-static void hashMapData_free(HashMapData data);
-
-static long long hash(const char *name);
-
-static long long hash1(double capacity, long long hash);
-
-static long long hash2(long long hash);
-
-static long long find(HashMap *map, const char *name, int flag);
-
-static int hashMap_put(HashMap *map, Function function, FuncType type);
-
-static void hashMap_free(HashMap *map);
-
-static int HashMap_remove(HashMap *map, char *name);
-
-
-
-
-/* #####################################################################################################################
- * AST functions.
- */
-
-/**
- * Abstract syntax tree used for arithmetic evaluations.
- */
-typedef struct AST AST;
-
-/**
- * Free the AST.
- * @param ast The AST pointer.
- * @return NULL.
- */
-static AST *AST_free(AST *ast);
-
-/**
- * Allocate a new memory block for the AST. its root is NULL.
- * @return The pointer of the AST.
- */
-static AST *AST_new();
-
-/**
- * Add a new Node in the AST according to the string given in argument. Manage the operator priority.
- * @param ast The AST.
- * @param storage. The storage ast.
- * @param s The node value.
- */
-static int AST_add(AST *ast, Tree *storage, Token token);
-
-/**
- * Apply the arithmetic expression stored in the ast.
- * @param storage The storage tree.
- * @param ast The AST.
- */
-static int AST_apply(Tree *storage, AST *ast, HashMap *map);
-
-/**
- *
- * @param ast
- * @return
- */
-static AST *AST_clear(AST *ast);
-
-/**
- * Abstract syntax node used for arithmetic evaluations.
- */
-typedef struct ASN ASN;
-
-/**
- * Allocate a new memory block for an ASN.
- * @param t The storage tree.
- * @param s The node's data
- * @return The node's pointer
- */
-static ASN *ASN_new(Tree *t, Token token);
-
-/**
- * Add a new Node in the AST according to the string given in argument. Manage the operator priority.
- * @param asn The root of the sub AST.
- * @param storage. The storage tree.
- * @param s The node value.
- */
-static int ASN_add(Tree *storage, ASN **asn, Token token);
-
-/**
- * Free the ASN.
- * @param asn The node which will be free.
- * @return NULL.
- */
-static ASN *ASN_free(ASN *asn);
-
-/**
- * Apply the arithmetic expression stored in the AST.
- * @param storage The storage tree.
- * @param asn The root of the sub AST.
- */
-static unbounded_int ASN_apply(Tree *storage, ASN *asn, int *err, HashMap *map);
-
-/**
- * Realize the operation represented by the node.
- *
- * @param asn The current node.
- * @param storage The storage tree.
- * @param left The left AST result.
- * @param right The right AST result
- * @return The value of the operation. <br>
- *         The DEFAULT_OP(empty node) is considered as an addition.<br>
- *         if the node is not an operator, return the result of the node.
- */
-static unbounded_int op(ASN *asn, Tree *storage, unbounded_int left, unbounded_int right);
-
-
-/* #####################################################################################################################
- * Parser functions.
- */
 /**
  * Open files according to the arguments given to the program. By default, in = stdin and out = stdout
  *
@@ -483,54 +202,10 @@ treatment(Buffer *pBuffer, AST *ast, Tree *storage, TokenType type, int *func, H
 
 static int function_apply(HashMap *map, char *name, ASN *node);
 
-/**
- * Indicate if the char given in argument is an operator
- * @param s The char which will be tested.
- * @return return true if the char is an operator, false otherwise.
- */
-static int isAnOperator(char s);
-
-/**
- * Indicate if the char a has a higher precedence than b.
- * @param a The first operator.
- * @param b The second operator.
- * @return true if a has a higher precedence than b, false otherwise.
- */
-static int isHigher(char a, char b);
-
-/**
- * Associate a priority with an operator.
- * @param a The operator.
- * @return The operator's priority
- */
-static unsigned short priority(char a);
-
-/**
- * Trim the string with the given regex. <br>
- * For example, all the space before the other character is ignored, and all the space after are deleted.
- * @param s The string which will be treated.
- * @return return the trimmed string
- */
-static char *trim(const char *s, size_t len);
-
-/**
- * Check is the given character is a sign symbol (+ or -) or is a number
- * @param c The character that will be tested.
- * @return true if its the case, false otherwise.
- */
-static int isSignOrNumber(char c);
-
-static char *intToString(int n);
-
-
 static Function buildCalledFunction(HashMapData data, char *buffer);
 
 static int functionTreatment(int c, Buffer *buffer, AST *ast, Tree *storage, int *isFunc, int *argsStart, HashMap *map,
                              Function *function);
-
-static int AST_hasFunction(AST *ast);
-
-static int ASN_hasFunction(ASN *asn);
 
 /* #####################################################################################################################
  * struct definitions
@@ -546,18 +221,6 @@ struct Node {
 struct Tree {
     Node *root;
 };
-
-struct ASN {
-    Token token;
-    unbounded_int result;
-    ASN *left;
-    ASN *right;
-};
-
-struct AST {
-    ASN *root;
-};
-
 
 void load_stdlib(HashMap *map) {
     Function print = function_new("print", VOID_TYPE, std_print, 1);
@@ -878,18 +541,6 @@ static int treatment
     return 1; //add value
 }
 
-static int AST_hasFunction(AST *ast) {
-    if (ast == NULL || ast->root == NULL) return 0;
-    return ASN_hasFunction(ast->root);
-}
-
-static int ASN_hasFunction(ASN *asn) {
-    if (asn->token.type == FUNCTION) return 1;
-    int left = (asn->left == NULL) ? 0 : ASN_hasFunction(asn->left);
-    int right = (asn->right == NULL) ? 0 : ASN_hasFunction(asn->right);
-    return left || right;
-}
-
 static Function buildCalledFunction(HashMapData data, char *buffer) {
     Function result = FUNCTION_NULL;
     int dLen = (int) strlen(data.function.name);
@@ -969,47 +620,6 @@ static void disconnect(FILE **in, FILE **out) {
         *out = NULL;
     }
 }
-
-static char *error_getMessage(enum ERROR error) {
-    switch (error) {
-        case INVALID_SYNTAX:
-            return "An invalid syntax was found";
-        case INVALID_OPERATOR:
-            return "An unknown operator was found";
-        case INTERNAL:
-            return "Internal error occurred during the execution";
-        case MISSING_BLANK:
-            return "A blank is missing";
-        case UNKNOWN_CHARACTER:
-            return "An Unknown character was found";
-        case TOO_MANY_ARGUMENTS:
-            return "Too many arguments were given";
-        case MISSING_ARGUMENTS:
-            return "Some arguments are missing";
-        default:
-            return "";
-    }
-}
-
-
-/* #####################################################################################################################
- * Token's functions.
- */
-
-static Token token_new(char *s, size_t len, enum TokenType type) {
-    Token token = {.data = trim(s, len), .type = type};
-    return token;
-}
-
-static void token_free(Token t) {
-    if (t.type == FUNCTION) return;
-    //free(t.data);
-}
-
-
-/* #####################################################################################################################
- * Parser buffer function
- */
 
 static Buffer *buffer_new() {
     Buffer *buffer = malloc(sizeof(Buffer));
@@ -1110,163 +720,6 @@ static int std_fact(int argc, unbounded_int *argv, char **argn) {
  * AST functions
  */
 
-static AST *AST_new() {
-    AST *tree = malloc(sizeof(AST));
-    if (tree == NULL) return NULL;
-    tree->root = NULL;
-    return tree;
-}
-
-static ASN *ASN_new(Tree *t, Token token) {
-    ASN *node = malloc(sizeof(ASN));
-    if (node == NULL) return NULL;
-    switch (token.type) {
-        case OPERATOR:
-            node->result = ll2unbounded_int(0);
-            node->token = token;
-            break;
-        case NUMBER:
-            node->token = token;
-            node->result = string2unbounded_int(token.data);
-            break;
-        case VAR:
-            node->result = tree_getValue(t, token.data);
-            node->token = token;
-            break;
-            node->token = token;
-        case FUNCTION:
-            node->result = ll2unbounded_int(0);
-
-        default:
-            node->result = ll2unbounded_int(0);
-            node->token = token;
-            break;
-    }
-
-    node->right = NULL;
-    node->left = NULL;
-    return node;
-}
-
-static int AST_add(AST *ast, Tree *storage, Token token) {
-    if (ast == NULL || storage == NULL || isspace((unsigned )token.data[0]) || token.type == VOID) {
-        pERROR(INTERNAL);
-        return 0;
-    }
-    return ASN_add(storage, &ast->root, token);
-}
-
-static int ASN_add(Tree *storage, ASN **asn, Token token) {
-    if (*asn == NULL) {
-        *asn = ASN_new(storage, token_new("", 0, OPERATOR));
-        (*asn)->left = ASN_new(storage, token);
-    } else if ((*asn)->token.data[0] == DEFAULT_OP) {
-        if ((*asn)->token.type != token.type) {
-            pERROR(INVALID_SYNTAX);
-            return 0;
-        }
-        strncpy((*asn)->token.data, token.data, strlen(token.data));
-    } else if ((*asn)->right == NULL || token.type == NUMBER || token.type == FUNCTION || token.type == VAR ||
-               !isHigher((*asn)->token.data[0], token.data[0])) {
-        return ASN_add(storage, &(*asn)->right, token);
-    } else {
-        ASN *tmp = *asn;
-        *asn = ASN_new(storage, token);
-        (*asn)->left = tmp;
-    }
-    return 1;
-}
-
-static AST *AST_clear(AST *ast) {
-    if (ast->root != NULL) {
-        ASN_free(ast->root);
-    }
-    ast->root = NULL;
-    return ast;
-}
-
-static AST *AST_free(AST *ast) {
-    ASN_free(ast->root);
-    free(ast);
-    return NULL;
-}
-
-static ASN *ASN_free(ASN *n) {
-    if (n != NULL) {
-        ASN_free(n->right);
-        ASN_free(n->left);
-        //unbounded_int_free(n->result);
-        free(n);
-    }
-    return n = NULL;
-}
-
-static unbounded_int op(ASN *asn, Tree *storage, unbounded_int left, unbounded_int right) {
-    switch (asn->token.data[0]) {
-        case '=':
-            EQUALS(storage, right, asn->left->token.data);
-            return right;
-        case '*':
-            return MULTIPLICATION(right, left);
-        case '-':
-            return SUBSTRACTION(left, right);
-        case DEFAULT_OP:
-        case '+':
-            return ADDITION(left, right);
-            /*case '/':
-                return DIVISION(left, right);
-                case '%':
-                    return MODULO(left, right);*/
-        default:
-            return asn->result;
-    }
-}
-
-static int AST_apply(Tree *storage, AST *ast, HashMap *map) {
-    if (storage == NULL || ast == NULL) {
-        errno = 22; //Invalid arguments
-        printErr("Internal Error");
-        return 0;
-    }
-    if (ast->root == NULL) {
-        pERROR(INVALID_SYNTAX);
-        return 0;
-    }
-    if (ast->root->token.data[0] == DEFAULT_OP && ast->root->left == NULL) {
-        pERROR(INVALID_OPERATOR);
-        return 0;
-    }
-    int err = 1;
-    ASN_apply(storage, ast->root, &err, map);
-    return err;
-}
-
-static unbounded_int ASN_apply(Tree *storage, ASN *asn, int *err, HashMap *map) {
-    if (asn == NULL) return ll2unbounded_int(0);
-    if (asn->token.type == NUMBER || asn->token.type == VAR) {
-        return asn->result;
-    }
-    unbounded_int left = ASN_apply(storage, asn->left, err, map);
-    unbounded_int right = ASN_apply(storage, asn->right, err, map);
-    if (((asn->token.data[0] != DEFAULT_OP) && asn->token.type == OPERATOR) && // missing binary operator argument
-        ((asn->left == NULL) || (asn->right == NULL))) {
-        if (*err) pERROR(INVALID_SYNTAX);
-        *err = 0;
-        return ll2unbounded_int(0);
-    }
-    if (asn->token.data[0] == '=' && asn->left->token.type == NUMBER) { //Invalid assignation
-        if (*err) pERROR(INVALID_SYNTAX);
-        *err = 0;
-        return ll2unbounded_int(0);
-    }
-    if (asn->token.type == FUNCTION) {
-        *err = function_apply(map, asn->token.data, asn);
-        return asn->result;
-    }
-    return asn->result = op(asn, storage, left, right);
-}
-
-
 /* #####################################################################################################################
  * Storage tree's functions.
  */
@@ -1354,7 +807,6 @@ static void node_free(Node *n) {
 /* #####################################################################################################################
  * Function's functions
  */
-
 static int function_apply(HashMap *map, char *name, ASN *node) {
     HashMapData data = hashMap_get(map, name);
     if (data.type == NONE) return 0;
@@ -1374,249 +826,12 @@ static int function_apply(HashMap *map, char *name, ASN *node) {
     return 1;
 }
 
-static Function
-function_new(char *name, RetType type, int (*function)(int, unbounded_int *, char **),
-             unsigned short requestedArguments) {
-    Function f = {.name = trim(name, strlen(name)),
-            .requested = requestedArguments,
-            .func = function,
-            .argc = 0,
-            .retType = type,
-            .argv = calloc(requestedArguments + 1, sizeof(unbounded_int)),
-            .argn = malloc(requestedArguments * sizeof(char *))};
-    return f;
-}
-
-static void function_free(Function f) {
-    //free(f.argv);
-    //free(f.argn);
-    free(f.name);
-}
-
 
 /* #####################################################################################################################
  * HashMap's function
  */
-static long long hash(const char *name) {
-    double value = 0;
-    double len = (double) strlen(name);
-    for (int i = 0; i < len; i++) {
-        value += (int) name[i] * (pow(31, len - i));
-    }
-    return (long long) value;
-}
-
-static long long hash1(double capacity, long long hash) {
-    return (long long) (((double) (hash) / HASH_A) * capacity);
-}
-
-static long long hash2(long long hash) {
-    return (long long) (2 * hash + 1) % LLONG_MAX;
-}
-
-static HashMap *HashMap_new() {
-    HashMap *map = malloc(sizeof(HashMap));
-    if (map == NULL) {
-        printErr("");
-        return NULL;
-    }
-    map->minRatio = HASHMAP_MIN_RATIO;
-    map->maxRatio = HASHMAP_MAX_RATIO;
-    map->capacity = HASHMAP_INITIAL_SIZE;
-    map->data = malloc(sizeof(HashMapData) * map->capacity);
-    for (int i = 0; i < map->capacity; i++) {
-        map->data[i] = NONE_DATA;
-    }
-    if (map->data == NULL) {
-        printErr("");
-        free(map);
-        return NULL;
-    }
-    map->keyNumber = 0;
-    map->dummyNumber = 0;
-    return map;
-}
-
-static long long find(HashMap *map, const char *name, int flag) {
-    long long hashVal = hash(name);
-    long long ind = hash1((double) map->capacity, hashVal) % (long long) map->capacity;
-    long long step = hash2(hashVal);
-    long long dummy = -1;
-    for (int i = 0; i < map->capacity; i++) {
-        if (map->data[ind].type == NONE) {
-            if (dummy == -1) dummy = ind;
-            break;
-        }
-        if (map->data[ind].type == DUMMY) {
-            if (dummy == -1) dummy = ind;
-        } else if (map->data[ind].hash == hashVal && str_equals(map->data[ind].function.name, name)) {
-            return ind;
-        }
-        ind = (ind + step) % (long long) map->capacity;
-    }
-    return (flag) ? dummy : -1;
-}
-
-static int hashMap_put(HashMap *map, Function function, FuncType type) {
-    long long hashVal = hash(function.name);
-    long long empty = find(map, function.name, 1);
-    if (empty == -1) return 0;
-
-    HashMapData new = {.function = function, .type = type, .hash = hashVal};
-    if (map->data[empty].type == DUMMY) {
-        map->dummyNumber--;
-        map->keyNumber++;
-    }
-    if (map->data[empty].type == NONE) {
-        map->keyNumber++;
-    }
-    map->data[empty] = new;
-    if (map->dummyNumber + map->keyNumber >= (long long) ((double) map->capacity * map->maxRatio)) {
-        int ratio = (map->dummyNumber < map->keyNumber) ? 2 : 1;
-        resize(map, ratio * map->capacity);
-    }
-    return 1;
-}
-
-static HashMapData hashMap_get(HashMap *map, const char *name) {
-    long long pos = find(map, name, 0);
-    if (pos == -1) return NONE_DATA;
-    return map->data[pos];
-}
-
-static void resize(HashMap *map, size_t newSize) {
-    HashMapData *old = map->data;
-    size_t oldLen = map->capacity;
-    HashMapData *new = malloc(newSize * sizeof(HashMapData));
-    for (int i = 0; i < newSize; i++) new[i] = NONE_DATA;
-    if (new == NULL) {
-        printErr("");
-        return;
-    }
-    map->data = new;
-    map->capacity = newSize;
-    map->keyNumber = 0;
-    map->dummyNumber = 0;
-    for (int i = 0; i < oldLen; i++) {
-        if (old[i].type == NONE || old[i].type == DUMMY) continue;
-        else hashMap_put(map, old[i].function, old[i].type);
-    }
-    free(old);
-}
-
-static void hashMapData_free(HashMapData data) {
-    if (data.type == NONE || data.type == DUMMY) return;
-    function_free(data.function);
-}
-
-static int HashMap_remove(HashMap *map, char *name) {
-    long long pos = find(map, name, 0);
-    if (pos == -1) return -1;
-    HashMapData del = map->data[pos];
-    map->data[pos] = DUMMY_DATA(del.hash);
-    hashMapData_free(del);
-    map->dummyNumber++;
-    map->keyNumber--;
-    if (map->dummyNumber + map->keyNumber <= (long long) ((double) map->capacity * map->minRatio)) {
-        resize(map, (map->capacity / 2));
-    }
-    return 1;
-}
-
-static void hashMap_free(HashMap *map) {
-    if (map == NULL) return;
-    for (int i = 0; i < map->capacity; i++) {
-        hashMapData_free(map->data[i]);
-    }
-    free(map);
-}
 
 
 /* #####################################################################################################################
  * Other's functions
  */
-static int isHigher(const char a, const char b) {
-    if (!isAnOperator(a)) return 0;
-    if (!isAnOperator(b)) return 1;
-    return priority(a) > priority(b);
-}
-
-static unsigned short priority(const char a) {
-    switch (a) {
-        case '=':
-            return 1;
-        case '+':
-        case '-':
-            return 2;
-            //case '/':
-        case '*':
-            //case '%':
-            return 3;
-        default:
-            return 0;
-    }
-}
-
-static char *trim(const char *s, size_t len) {
-    int start = 0, end = 0;
-    for (int i = 0; i < (int) len; i++) {
-        if (!isspace((unsigned) s[i])) {
-            start = i;
-            break;
-        }
-    }
-    for (int i = (int) len; i >= 0; i--) {
-        if (!isspace((unsigned) s[i])) {
-            end = i;
-            break;
-        }
-    }
-    int size = end - start;
-    char *ret = malloc((size) * sizeof(char) + 1);
-    if (ret == NULL) {
-        printErr("");
-        return NULL;
-    }
-    strncpy(ret, s, size);
-    ret [size] = '\0';
-    return ret;
-}
-
-static int isSignOrNumber(char c) {
-    return c == '+' || c == '-' || isdigit((unsigned)c);
-}
-
-static int nDigits(unsigned int i) {
-    int n = 1;
-    while (i > 9) {
-        n++;
-        i /= 10;
-    }
-    return n;
-}
-
-static char *intToString(int n) {
-    int len = nDigits(abs(n));
-    char *s = malloc(len * sizeof(char));
-    if (s == NULL) return NULL;
-    snprintf(s, len + 1, "%d", n);
-    return s;
-}
-
-static int str_equals(const char *s1, const char *s2) {
-    int l1 = (int) strlen(s1);
-    int l2 = (int) strlen(s2);
-    if (l1 != l2) return 0;
-    for (int i = 0; i < l1; i++) {
-        if (s1[i] != s2[i]) {
-            return 0;
-        }
-    }
-    return 1;
-}
-
-static int isAnOperator(const char s) {
-    return s == '*' || s == '+' || s == '-' || s == '=';
-    //|| s == '%' || s == '/'
-
-}
