+++ src/stack.h
@@ -1,0 +1,8 @@
+//
+// Created by Kevin on 05/07/2022.
+//
+
+#ifndef C_PROJECT_STACK_H
+#define C_PROJECT_STACK_H
+
+#endif //C_PROJECT_STACK_H
