+++ src/statck.c
@@ -1,0 +1,5 @@
+//
+// Created by Kevin on 05/07/2022.
+//
+
+#include "stack.h"
