+++ test/test_unbounded.c
@@ -2,7 +2,7 @@
 #include<stdlib.h>
 #include<string.h>
 
-#include "unbounded_int.h"
+#include "src/unbounded_int.h"
 
 /**
  * fonction main (tests)
